requirements.txt
 fastapi 
 uvicorn 
 google-generativeai 
 langchain 
 langchain-google-genai 
 chromadb 
 tavily-python
